PART 4 FINISHED

+++ part4/bloglist/controllers/blogRouter.js
@@ -1,32 +1,73 @@
 const blogRouter = require('express').Router()
-const { request } = require('../app')
-const app = require('../app')
-const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 const logger = require('../utils/logger')
+const jwt = require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response, next) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user')
   response.json(blogs)
 })
 
 blogRouter.post('/', async (request, response, next) => {
-  const newBlog = request.body
+  const body = request.body
 
-  if (newBlog.title === undefined) {
-    return response.status(400).json({ error: 'Missing title'})
-  } else if (newBlog.url === undefined) {
-    return response.status(400).json({ error: 'Missing url'})
+  const token = request.token
+  // jwt.verify throws UnhandledRejectedPromises everywhere
+  // so use try - catch and take care manually
+  try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (body.title === undefined) {
+      return response.status(400).json({ error: 'Missing title'})
+    } else if (body.url === undefined) {
+      return response.status(400).json({ error: 'Missing url'})
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = new Blog({
+      title: body.title,
+      url: body.url,
+      likes: body.likes,
+      author: body.author,
+      user: user._id
+    })
+
+    const postedBlog = await blog.save()
+    user.blogs = user.blogs.concat(postedBlog._id)
+    await user.save()
+    response.status(200).json(postedBlog)
+    
+  } catch(error) {
+    error.name = "JsonWebTokenError"
+    next(error)
   }
-
-  const blog = new Blog(request.body)
-  const postedBlog = await blog.save()
-  response.status(200).json(postedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  // jwt.verify throws UnhandledRejectedPromises everywhere
+  // so use try - catch and take care manually
+  try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (user._id.toString() !== blog.user.toString()) {
+      return response.status(401).json({ error: 'only the user that made the entry can delete it'})
+    }
+    await Blog.findByIdAndDelete(request.params.id)
+    user.blogs = user.blogs
+      .filter(id => id.toString() !== request.params.id)
+    await user.save()
+    response.status(204).end()
+  } catch(error) {
+    error.name = "JsonWebTokenError"
+    next(error)
+  }
+  
+  
+  
+
+  
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
